Removed MongoDb after testing it out, using SQLite until a better replacement shows up. changed sorting an members to a more DRY one.

+++ Models/MembershipType.cs
@@ -2,7 +2,6 @@ using System.ComponentModel.DataAnnotations;
 
 namespace SPIIKcom.Models
 {
-	// Add profile data for application users by adding properties to the ApplicationUser class
 	public class MembershipType
 	{
 		public int ID { get; set; }
@@ -10,7 +9,7 @@ namespace SPIIKcom.Models
 		public string Name { get; set; }
 		[Required]
 		public decimal Price { get; set; }
-		[Required]		
-		public double LengthInYears { get; set; }
+		[Required]
+		public int LengthInYears { get; set; }
 	}
 }
